PFM-ISSUE-28877: Integrate additional output into the merge process.
